major refactoring - deprecated classes and methods removed
