+++ src/main/java/sudoku/readers/FileSudokuReader.java
@@ -1,4 +1,4 @@
-package sudoku.processing;
+package sudoku.readers;
 
 import java.io.BufferedReader;
 import java.io.File;
