+++ src/main/resources/inputs/extremelyHard.txt
@@ -1,0 +1,9 @@
+0 0 2 0 0 3 8 5 6
+9 0 0 4 0 0 0 0 0
+0 0 7 0 0 5 0 0 0
+0 4 0 1 2 0 0 0 7
+0 0 0 0 0 0 0 0 0
+5 0 0 0 6 4 0 3 0
+0 0 0 3 0 0 7 0 0
+0 0 0 0 0 9 0 0 4
+7 3 6 8 0 0 9 0 0
