+++ src/main/resources/outputs/extremelyHard.txt
@@ -1,0 +1,9 @@
+4 1 2 7 9 3 8 5 6
+9 8 5 4 1 6 2 7 3
+3 6 7 2 8 5 4 1 9
+6 4 3 1 2 8 5 9 7
+1 2 9 5 3 7 6 4 8
+5 7 8 9 6 4 1 3 2
+8 9 4 3 5 2 7 6 1
+2 5 1 6 7 9 3 8 4
+7 3 6 8 4 1 9 2 5
