+++ src/test/java/SudokuTest.java
@@ -6,7 +6,7 @@ import sudoku.model.Box;
 import sudoku.model.Column;
 import sudoku.model.Row;
 import sudoku.model.Sudoku;
-import sudoku.processing.FileSudokuReader;
+import sudoku.readers.FileSudokuReader;
 
 import java.io.File;
 import java.util.List;
@@ -17,7 +17,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 public class SudokuTest {
 
-    static ClassLoader classLoader = new SolutionTest().getClass().getClassLoader();
+    static ClassLoader classLoader = new SudokuTest().getClass().getClassLoader();
 
     private static final String inp1 = new File(classLoader.getResource("inputs/simple1.txt").getFile()).getPath();
     private static final String inp2 = new File(classLoader.getResource("inputs/simple2.txt").getFile()).getPath();
