+++ src/test/java/sudoku/processingUsingStrategy/HiddenSingleInACellTest.java
@@ -1,12 +1,11 @@
 package sudoku.processingUsingStrategy;
 
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import sudoku.customExceptions.IllegalSudokuStateException;
 import sudoku.model.Sudoku;
-import sudoku.processing.FileSudokuReader;
+import sudoku.readers.FileSudokuReader;
 
 import java.io.File;
 import java.util.stream.Stream;
