+++ src/test/java/sudoku/processingUsingStrategy/NakedSingleInACellTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import sudoku.customExceptions.IllegalSudokuStateException;
 import sudoku.model.Sudoku;
-import sudoku.processing.FileSudokuReader;
+import sudoku.readers.FileSudokuReader;
 
 import java.io.File;
 import java.util.stream.Stream;
