+++ src/test/java/sudoku/processingUsingStrategy/PointingPairsInCellTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import sudoku.customExceptions.IllegalSudokuStateException;
 import sudoku.model.Sudoku;
-import sudoku.processing.FileSudokuReader;
+import sudoku.readers.FileSudokuReader;
 
 import java.io.File;
 import java.util.stream.Stream;
